refactor : show modal from calendar view
