+++ forever_planner/src/views/CalendarView.vue
@@ -1,4 +1,6 @@
 <template>
+  <DateModal v-if="dateModalState" />
+  <CategoryModal v-if="categoryModalState" />
   <div class="w-full h-screen flex flex-col justify-between">
     <div class="w-full h-full bg-[#f5f7fd] p-2 flex flex-col justify-end">
       <div class="text-2xl font-semibold p-3">7월</div>
@@ -11,15 +13,26 @@
 <script setup>
 import CalendarVue from '@/components/Calendar/CalendarVue.vue'
 import FooterVue from '@/components/FooterVue.vue';
+import DateModal from '@/components/Calendar/DateModal.vue';
+import CategoryModal from '@/components/Calendar/Category/CategoryModal.vue';
 import { storeToRefs } from 'pinia';
-import { useStore } from '@/stores/store.js';
-import { onMounted } from 'vue';
+import { useModalStore } from '@/stores/modalStore.js';
+import { onMounted, watchEffect } from 'vue';
 
-const { isClicked } = storeToRefs(useStore());
+const { dateModalState, categoryModalState } = storeToRefs(useModalStore());
+
+const handleStopScroll = () => {
+  if (dateModalState.value) document.documentElement.style.overflow = 'hidden';
+  else document.documentElement.style.overflow = 'auto';
+};
+
+watchEffect(() => {
+  handleStopScroll();
+});
 
 onMounted(() => {
-  isClicked.value = 'diary'
-})
+  window.scrollTo(0, 0);
+});
 </script>
 
 <style scoped>
